Refactor: Enhance sign-in feature and remove unused screens
This commit refactors the sign-in functionality and removes the unused `ForgetPassword` and `SignUpScreen` files.

Key changes include:
- Updated `AuthRemoteDataSourceImpl` to use `SigninRequestEntity` and `SigninResponseEntity`.
- Added mock for `ApiServices` for testing purposes.
- Renamed `SigninResponseEntity` to `UserResponseEntity` and created a new `SigninResponseEntity`.
- Updated `SigninRequestDto` to use non-nullable types and added `toDto` method in `SigninRequestEntity`.
- Separated `UserDto` from `SigninResponseDto` into its own file.
- Updated `AuthRemoteDataSource` to use `SigninRequestEntity`.
- Updated `AuthRepoImpl` to use `SigninRequestEntity`.
- Created `UserDto` to represent user data transfer object.
- Updated `AuthRepo` to use `SigninRequestEntity`.
- Updated `SigninUseCase` to use `SigninRequestEntity`.
- Created `SigninRequestEntity` to represent sign-in request data.
- Updated `SigninViewModel` to manage form state and use `SigninRequestEntity`.
- Updated `SigninScreen` to use `SigninViewModel` for form handling and state management, and integrated validation logic.
- Removed `ForgetPassword` screen.
- Removed `SignUpScreen`.
