+++ lib/Feature/auth/api/model/requests/sign_in_request_dto.dart
@@ -1,17 +1,16 @@
 import 'package:json_annotation/json_annotation.dart';
-
 part 'sign_in_request_dto.g.dart';
 
 @JsonSerializable()
 class SigninRequestDto {
   @JsonKey(name: "email")
-  final String? email;
+  final String email;
   @JsonKey(name: "password")
-  final String? password;
+  final String password;
 
   SigninRequestDto ({
-    this.email,
-    this.password,
+    required this.email,
+    required this.password,
   });
 
   factory SigninRequestDto.fromJson(Map<String, dynamic> json) {
