+++ lib/Feature/auth/api/model/responses/sign_in_response_dto.dart
@@ -1,4 +1,5 @@
-import 'package:flower_e_commerce_app/Feature/auth/domain/Entity/sign_in_entity.dart';
+import 'package:flower_e_commerce_app/Feature/auth/api/model/responses/user_response_dto.dart';
+import 'package:flower_e_commerce_app/Feature/auth/domain/entity/response/sign_in_response_entity.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'sign_in_response_dto.g.dart';
@@ -8,7 +9,7 @@ class SigninResponseDto {
   @JsonKey(name: "message")
   final String? message;
   @JsonKey(name: "user")
-  final User? user;
+  final UserDto? user;
   @JsonKey(name: "token")
   final String? token;
 
@@ -34,69 +35,3 @@ class SigninResponseDto {
     );
   }
 }
-
-@JsonSerializable()
-class User {
-  @JsonKey(name: "_id")
-  final String? Id;
-  @JsonKey(name: "firstName")
-  final String? firstName;
-  @JsonKey(name: "lastName")
-  final String? lastName;
-  @JsonKey(name: "email")
-  final String? email;
-  @JsonKey(name: "gender")
-  final String? gender;
-  @JsonKey(name: "phone")
-  final String? phone;
-  @JsonKey(name: "photo")
-  final String? photo;
-  @JsonKey(name: "role")
-  final String? role;
-  @JsonKey(name: "wishlist")
-  final List<dynamic>? wishlist;
-  @JsonKey(name: "addresses")
-  final List<dynamic>? addresses;
-  @JsonKey(name: "createdAt")
-  final String? createdAt;
-
-  User ({
-    this.Id,
-    this.firstName,
-    this.lastName,
-    this.email,
-    this.gender,
-    this.phone,
-    this.photo,
-    this.role,
-    this.wishlist,
-    this.addresses,
-    this.createdAt,
-  });
-
-  factory User.fromJson(Map<String, dynamic> json) {
-    return _$UserFromJson(json);
-  }
-
-  Map<String, dynamic> toJson() {
-    return _$UserToJson(this);
-  }
-
-  UserEntity toEntity() {
-    return UserEntity(
-      Id: Id,
-      firstName: firstName,
-      lastName: lastName,
-      email: email,
-      gender: gender,
-      phone: phone,
-      photo: photo,
-      role: role,
-      wishlist: wishlist,
-      addresses: addresses,
-      createdAt: createdAt,
-    );
-  }
-}
-
-
