+++ lib/Feature/auth/domain/repositories/auth_repo.dart
@@ -1,10 +1,10 @@
-import 'package:flower_e_commerce_app/Feature/auth/domain/Entity/sign_in_entity.dart';
 import 'package:flower_e_commerce_app/core/Errors/api_results.dart';
+import '../entity/request/sign_in_request_entity.dart';
+import '../entity/response/sign_in_response_entity.dart';
 
 abstract interface class AuthRepo {
   Future<ApiResult<SigninResponseEntity>> signin({
-    required String email,
-    required String password,
+    required SigninRequestEntity request,
     bool? rememberMeChecked
   });
 }
