--- lib/Feature/auth/presentation/pages/forget_password.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ForgetPassword extends StatefulWidget {
-  const ForgetPassword({super.key});
-
-  @override
-  State<ForgetPassword> createState() => _ForgetPasswordState();
-}
-
-class _ForgetPasswordState extends State<ForgetPassword> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Forget Password')),
-      body: const Center(child: Text('Forget Password')),
-    );
-  }
-}
