--- lib/Feature/auth/presentation/pages/sign_up_screen.dart
@@ -1,19 +1,0 @@
-import 'package:flower_e_commerce_app/core/Widgets/custom_app_bar.dart';
-import 'package:flutter/material.dart';
-
-class SignUpScreen extends StatefulWidget {
-  const SignUpScreen({super.key});
-
-  @override
-  State<SignUpScreen> createState() => _SignUpScreenState();
-}
-
-class _SignUpScreenState extends State<SignUpScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Sign Up')),
-      body: const Center(child: Text('Sign Up')),
-    );
-  }
-}
