+++ test/mocks/api_services.dart
@@ -1,0 +1,5 @@
+import 'package:mockito/annotations.dart';
+import 'package:flower_e_commerce_app/Feature/auth/api/client/api_service.dart';
+
+@GenerateMocks([ApiServices])
+void main() {}
